Fixing scripts, again.

+++ scripts/createService.sh
@@ -1,1 +1,4 @@
-ln -s /opt/id-service/*.jar /etc/init.d/id-service
+if [ -f /etc/init.d/id-service ];
+then
+    ln -s /opt/id-service/*.jar /etc/init.d/id-service
+fi
